Move ConsoleErrorHandler to Error
Having this class in the Error package makes the splits for Console
libraries much simpler as we don't need a split for Routing and Error.
